fixed delete_folder func.

+++ src/utils/utils.py
@@ -30,9 +30,10 @@ def gene_dict_to_table(rpkm_gene, sample_name):
     return df
 
 def delete_folder(top):
-    for root, dirs, files in os.walk(top, topdown=False):
-        for name in files:
-            os.remove(os.path.join(root, name))
-        for name in dirs:
-            os.rmdir(os.path.join(root, name))
-    os.rmdir(top)
+    if os.path.isdir(top):
+        for root, dirs, files in os.walk(top, topdown=False):
+            for name in files:
+                os.remove(os.path.join(root, name))
+            for name in dirs:
+                os.rmdir(os.path.join(root, name))
+        os.rmdir(top)
